Prevents multiple file download when range reading unsupported on server
If range reading was unsupported, but the viewer still requested it, the nifti file
ended up being downloaded twice, as it first was requested (and cached) as header
only, and then, when it was time to show it, it was requested and fully downloaded
again.

This solution basically copies the cached file fetch promise from the first bytes only
fetcher to the whole file fetcher in case we requested only the initial bytes, but
the server sent it whole.